+++ examples/simple.rs
@@ -1,15 +1,15 @@
-use line_numbers::NewlinePositions;
+use line_numbers::LinePositions;
 
 fn main() {
     let s = "foo\nbar\nbaz\n";
     let s_lines: Vec<_> = s.lines().collect();
 
-    let newlines = NewlinePositions::from(s);
+    let line_positions = LinePositions::from(s);
 
     let offset = 5;
-    let line_num = newlines.from_offset(offset);
+    let line_num = line_positions.from_offset(offset);
     println!(
-        "Offset {} is on line {}, which has text {:?}",
+        "Offset {} is on line {}, which has the text {:?}.",
         offset,
         line_num.display(),
         s_lines[line_num.as_usize()]