+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Replaced `LineNumber::one_indexed()` with `LineNumber::display()`, as
 one-indexed lines are only really for human consumption.
 
+Added function `NewlinePositions::from_offset()`.
+
 # 0.1.0 (released 5th August 2023)
 
 Initial release.