+++ CHANGELOG.md
@@ -5,6 +5,8 @@ one-indexed lines are only really for human consumption.
 
 Added function `NewlinePositions::from_offset()`.
 
+Renamed `NewlinePositions` to `LinePositions`.
+
 # 0.1.0 (released 5th August 2023)
 
 Initial release.